Refactor: Temporarily disable PreviewParameter in PokemonScreen previews
This commit comments out the `@PreviewParameter` annotation in all `@Preview` composables within `PokemonScreen.kt`.
The `PokemonParamProvider` and its associated `pokemon` constant are moved to the bottom of the file.

This change is likely temporary to address a build or preview issue, as the preview functions now lack specific data to render.